+++ src/components/switch/switch-interface.ts
@@ -1,3 +1,6 @@
 export interface SwitchChangeEventDetail {
     checked: boolean;
-  }
+}
+
+export type SwitchLabelAlignment = 'top' | 'center' | 'bottom';
+